[1] move out operation for files partially works
